PE-5845 NC - Creating a group with diacritical marks (áéíóúäëïöüñçâê) not displaying properly
There was (slurp) call which converts stream resource to Java string.
And somewhere in this conversion diacritics is lost. This patch makes it so that
request body is buffered in a stream and passed directly to output.

This PR needs to be merged together with clj-http-request PR. See corresponding ticket.