(maint) Update require & import to match current conventions
The previous usage of referencing fully qualified java class(es) and not
namespacing libraries does not match current puppet clj best practices.

Additional changes are about to be made in service of TK-214 and
updating the small amount of existing code to be in a similar style to
the code that will come from TK-214 should allow for easier code revieww.