(maint) Update dependencies and exclusions
Previously, we did not explicitly depend on trapperkeeper and
kitchensink at runtime. We will need to in the future and in adding
those projects to our runtime dependencies causes a cascade of
conflicts requiring fixes by both updating our exclusions and the
dependent versions we pull in.

This updates us to use the latest of all our dependencies, save for
clojure itself, in which we use 1.7.0 the version used by puppetserver.