(TK-214) Add tests for consolidated middleware and helpers
In the initial work for TK-214 we pulled in helpers from several
downstream products/libraries.  In that work there were few dedicated
tests to bring in as well (most tests that exercised these functions
were integration tests concerning general error handling by the
downstream app).

This adds tests around the behaviors of extracted functions.

Note: it also moves dev logging to be at the trace level so we can
capture all the output for the functions under test.