(maint) Work around JVM vs travis-ci hostname resolution
Travis-CI's execution environment and the JVM do not play well with each
other wrt hostname resolution.  Loading the JVM networking libraries (as
we do when we test middleware) will cause a buffer overflow in OpenJDK7
(I believe this is specifically an issue with the ubuntu package and how
travis has hostname/networking set up on their boxes).

This patch adds a pre-suite script to update the hostname and /etc/hosts
with the runtime hostname.  This is the recommended solution in
associated ticket: https://github.com/travis-ci/travis-ci/issues/5227