(maint) Move helpers to utils namespace
Previously we included middleware and utility methods all in the core
namespace.

This was awkward for consumers as they often only need one or the other.

This duplicates out all of the public utility functions into a utils
namespace, sets the core functions to depend on the them, and updates
the test accordingly. It leaves the existing implementations in the core
namespace so that consumers can upgrade components without having to
move in lockstep.