(SERVER-2344) Log stacktrace when handling uncaught Ex.
Previously we using the documented clojure.tool.logging interface for
printing Exception info. However it appears to not play nicely with our
logging configuration and does not print our stacktraces to a standard
location.

There may be a custom configuration option that could be toggled in
packaging, however, a more explicit and discoverable fix from
Engineering's side would be to explicitly gather the infromation we want
and directly print that.

This patch does that, gathering the stacktrace and string representation
and printing those rather than passing the exception into the logging
tools.

This also updates the test to ensure we do not send the stacktrace over
the wire.