(maint) Add params functions
This commit adds an implementation of the `params-request` function, based on
the implementations from upstream. This file was copied directly from the
puppetserver repo. It makes sense to keep all of our ring middleware utils in a
single place, both so there's a consistent place to find them and so that all
clojure components can make use of them.