(PE-30229) update caching headers
This updates the `wrap-add-cache-headers` middleware so that it adds a `cache-control` header
with the "no-store" directive instead of the previously-used "private, max-age=0, no-cache" directives.

The previous directives are out of date, as the "no-store" directive is now the only directive needed
to prevent cached responses on modern browsers.

Tests, the README, and the CHANGELOG are all updated to account for this change and prepare for a new release.