(maint) optimize wrap-uncaught-errors
This updates `wrap-uncaught-errors` to do several things.  First,
it guarantees that both the bye-output-stream, and the print-stream
used are properly closed.  Second, it moves the anonymous function
out of the wrapper into a separate declartion to avoid the cost of
creating the anonymous function with every instance of the wrapper.
It removes the use of `try+` as there is nothing that is
using the extended catch capabilities.

Finally, it updates the catch clause to catch Throwables, and not
just Exceptions.  This will allow it to catch Errors as well as
standard exceptions and return expected payloads.